Use hardcoded error history.
Now that we've completed the TS migration, we can clean up all the
migration related code/files from the support-frontend repo.
Currently this dashboard is requesting the data directly from the
github repo, but given the error history is now unchanging, we can
just move a hardcoded copy over to this repo for posterity.